backend/requirements.txt
Flask==2.3.3
python-telegram-bot==13.7
python-dotenv==0.19.0
flask-cors==3.0.10
Werkzeug==2.3.7